added config for pubsub for jmsTemplate

+++ web-app/main/java/com/scb/s2bx/nextgen/service/uiSolace/Publishable.java
@@ -2,6 +2,6 @@ package com.scb.s2bx.nextgen.service.uiSolace;
 
 public interface Publishable {
 
-    void next(int price);
+    void next(double price);
 
 }
